Move ApidocDefaults to _shared
